Split and reorganize test cases into separate files
Moved tests from Rejigs.Tests.cs into dedicated files: Alterations.cs, Building.cs, ComplexPatterns.cs, and Grouping.cs. This improves test organization and maintainability by grouping related tests together.
